set: 기본 패키지 및 환경 셋팅
