+++ src/main/java/com/hhplus/io/TicketReservationApplication.java
@@ -1,0 +1,13 @@
+package com.hhplus.io;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TicketReservationApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TicketReservationApplication.class, args);
+    }
+
+}
