+++ src/main/java/com/hhplus/io/common/config/SwaggerConfig.java
@@ -1,0 +1,20 @@
+package com.hhplus.io.common.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI openAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("콘서트 티켓 예매 시스템")
+                        .description("콘서트 티켓 예매 API")
+                        .version("1.0.0"));
+    }
+
+}
