+++ src/main/java/com/hhplus/io/concert/domain/Concert.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.concert.domain;
+
+public class Concert {
+}
