+++ src/main/java/com/hhplus/io/concert/persistence/ConcertJpaRepository.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.concert.persistence;
+
+public interface ConcertJpaRepository {
+}
