+++ src/main/java/com/hhplus/io/concert/persistence/ConcertRepositoryImpl.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.concert.persistence;
+
+public class ConcertRepositoryImpl {
+}
