+++ src/main/java/com/hhplus/io/concert/web/ConcertController.java
@@ -1,0 +1,8 @@
+package com.hhplus.io.concert.web;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ConcertController {
+
+}
