+++ src/main/java/com/hhplus/io/reservation/application/ReservationService.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.reservation.application;
+
+public class ReservationService {
+}
