+++ src/main/java/com/hhplus/io/reservation/domain/Reservation.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.reservation.domain;
+
+public class Reservation {
+}
