+++ src/main/java/com/hhplus/io/reservation/persistence/ReservationRepository.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.reservation.persistence;
+
+public interface ReservationRepository {
+}
