+++ src/main/java/com/hhplus/io/reservation/web/ReservationController.java
@@ -1,0 +1,7 @@
+package com.hhplus.io.reservation.web;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ReservationController {
+}
