+++ src/main/java/com/hhplus/io/user/application/UserService.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.user.application;
+
+public class UserService {
+}
