+++ src/main/java/com/hhplus/io/user/domain/User.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.user.domain;
+
+public class User {
+}
