+++ src/main/java/com/hhplus/io/user/persistence/UserRepository.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.user.persistence;
+
+public interface UserRepository {
+}
