+++ src/main/java/com/hhplus/io/user/web/UserController.java
@@ -1,0 +1,4 @@
+package com.hhplus.io.user.web;
+
+public class UserController {
+}
