--- src/main/java/com/hhplus/ticket/TicketReservationApplication.java
@@ -1,13 +1,0 @@
-package com.hhplus.ticket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TicketReservationApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TicketReservationApplication.class, args);
-    }
-
-}
