--- src/main/java/com/hhplus/ticket/common/config/SwaggerConfig.java
@@ -1,23 +1,0 @@
-package com.hhplus.ticket.common.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("콘서트 티켓 예매 시스템")
-                        .description("콘서트 티켓 예매 API")
-                        .version("1.0.0"));
-    }
-
-}
