--- src/main/java/com/hhplus/ticket/concert/application/ConcertService.java
@@ -1,7 +1,0 @@
-package com.hhplus.ticket.concert.application;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ConcertService {
-}
