--- src/main/java/com/hhplus/ticket/concert/domain/Concert.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.concert.domain;
-
-public class Concert {
-}
