--- src/main/java/com/hhplus/ticket/concert/persistence/ConcertJpaRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.concert.persistence;
-
-public interface ConcertJpaRepository {
-}
