--- src/main/java/com/hhplus/ticket/concert/persistence/ConcertRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.concert.persistence;
-
-public interface ConcertRepository {
-}
