--- src/main/java/com/hhplus/ticket/concert/persistence/ConcertRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.concert.persistence;
-
-public class ConcertRepositoryImpl {
-}
