--- src/main/java/com/hhplus/ticket/concert/web/ConcertController.java
@@ -1,8 +1,0 @@
-package com.hhplus.ticket.concert.web;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ConcertController {
-
-}
