--- src/main/java/com/hhplus/ticket/reservation/application/ReservationService.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.reservation.application;
-
-public class ReservationService {
-}
