--- src/main/java/com/hhplus/ticket/reservation/domain/Reservation.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.reservation.domain;
-
-public class Reservation {
-}
