--- src/main/java/com/hhplus/ticket/reservation/persistence/ReservationRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.reservation.persistence;
-
-public interface ReservationRepository {
-}
