--- src/main/java/com/hhplus/ticket/reservation/web/ReservationController.java
@@ -1,7 +1,0 @@
-package com.hhplus.ticket.reservation.web;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ReservationController {
-}
