--- src/main/java/com/hhplus/ticket/user/application/UserService.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.application;
-
-public class UserService {
-}
