--- src/main/java/com/hhplus/ticket/user/application/UserTokenFacade.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.application;
-
-public class UserTokenFacade {
-}
