--- src/main/java/com/hhplus/ticket/user/domain/UserToken.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.domain;
-
-public class UserToken {
-}
