--- src/main/java/com/hhplus/ticket/user/persistence/UserRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.persistence;
-
-public interface UserRepository {
-}
