--- src/main/java/com/hhplus/ticket/user/persistence/UserTokenRepository.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.persistence;
-
-public interface UserTokenRepository {
-}
