--- src/main/java/com/hhplus/ticket/user/web/UserController.java
@@ -1,4 +1,0 @@
-package com.hhplus.ticket.user.web;
-
-public class UserController {
-}
