+++ src/test/java/com/hhplus/io/TicketReservationApplicationTests.java
@@ -1,0 +1,13 @@
+package com.hhplus.io;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TicketReservationApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
