--- src/test/java/com/hhplus/ticket/TicketReservationApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hhplus.ticket;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TicketReservationApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
